Updated ReadTheDocs config for new structure
Added inherited members to the widgets' docs
Added missing requirement in requirements.txt
