+++ docs/api.rst
@@ -1,8 +1,17 @@
+API Definition
+---------------------
 
 .. If you created a package, create one automodule per module in the package.
 
 .. If your library file(s) are nested in a directory (e.g. /adafruit_foo/foo.py)
 .. use this format as the module name: "adafruit_foo.foo"
 
-.. automodule:: adafruit_progressbar
+.. automodule:: adafruit_progressbar.progressbar
    :members:
+   :show-inheritance:
+   :inherited-members:
+
+.. automodule:: adafruit_progressbar.verticalprogressbar
+   :members:
+   :show-inheritance:
+   :inherited-members:
